refactor: curry the contact reducer function to improve testability
