+++ etc/fetchers/fetch-sheets.ts
@@ -2,14 +2,8 @@ import fs from "fs";
 import path from "path";
 import cheerio from "cheerio";
 import fetch from "cross-fetch";
-import {
-  allIsEmptyString,
-  getKebabCase,
-  stripTags,
-  toSnakeCase,
-  toTitleCase,
-} from "../../lib/string-utils";
-import { extractGoogleQuery } from "./utils";
+import { allIsEmptyString, getKebabCase } from "../../lib/string-utils";
+import { contactReducer } from "./utils";
 
 export async function fetchSheets() {
   const source = await fetch("https://kcov.id/wbw-sheets");
@@ -62,33 +56,9 @@ export async function fetchSheets() {
         name: sheetName,
         slug: getKebabCase(sheetName),
         data: sheetRows.map((row, rowIndex) => {
-          return sheetColumns.reduce(
-            (prev: Record<string, number | string>, col) => {
-              const colName = toSnakeCase(col.name);
-              let cellValue = row[col.index];
-              if (colName == "lokasi") {
-                cellValue = toTitleCase(cellValue);
-              } else if (["kontak", "link"].includes(colName)) {
-                cellValue = extractGoogleQuery(cellValue);
-              } else if (colName == "ketersediaan") {
-                cellValue = toTitleCase(cellValue);
-              }
-              prev[colName] = cellValue;
-              if (colName == "kontak") {
-                prev.slug = [
-                  getKebabCase(prev.kebutuhan as string),
-                  getKebabCase(prev.keterangan as string),
-                  getKebabCase(prev.lokasi as string),
-                  getKebabCase(prev.penyedia as string),
-                  getKebabCase(stripTags(prev.kontak as string)),
-                ].join("-");
-              } else if (colName == "terakhir_update") {
-                prev.verifikasi = cellValue == "" ? 0 : 1;
-              }
-              return prev;
-            },
-            { id: rowIndex.toString() },
-          );
+          return sheetColumns.reduce(contactReducer(row), {
+            id: rowIndex.toString(),
+          });
         }),
       };
     })
